success: add action entry for a codified SimpleAction with Option<Variant> parametr, detect key pressed, get Control from it, obtain Action from Control, codify into a GioAction with Option<&Variant> parameter, activate SimpleAction with parameter, capture it in controller

+++ src/gui/action/gio_action_ty.rs
@@ -97,7 +97,7 @@ impl From<Action> for GioActionTy {
             Action::QuitDirectory => {
                 GioActionTy::new("quit-directory", GioActionParameterType::None)
             }
-            Action::Rank(_) => GioActionTy::new("rank", GioActionParameterType::Int32),
+            Action::Rank(_) => GioActionTy::new("rank", GioActionParameterType::Int64),
             Action::RemoveCategory(_) => {
                 GioActionTy::new("remove-category", GioActionParameterType::String)
             }
